other/sample_deploy/requirements.txt

# Function dependencies, for example:
# package>=version
# google-auth==1.22.1
# google-auth-httplib2==0.0.4
# google-auth-oauthlib==0.4.1
# google-cloud==0.34.0
# google-cloud-secret-manager==2.0.0
# gspread==3.6.0
# json5==0.9.5
# oauth2client==4.1.3
# oauthlib==3.1.0
# pandas==1.0.4
# pandas-gbq==0.14.0

flask==2.2.2
functions_framework==3.2.0
pandas
datetime
pandas_gbq
google-auth
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-secret-manager==2.0.0